changed pcap header fields back to uint32_t

+++ include/hdlc_parser/writer/pcap/pcap_header.hpp
@@ -16,8 +16,8 @@ struct PcapGlobalHeader {
 struct PcapPacketHeader {
     uint32_t ts_sec;
     uint32_t ts_usec;
-    size_t incl_len;
-    size_t orig_len;
+    uint32_t incl_len;
+    uint32_t orig_len;
 };
 
 #pragma pack(pop)
